# specialize simplelru(span's lruMap) for attributes only