… example/basic/go.sum
… example/http-stackdriver/go.sum
… example/http/go.sum
… example/jaeger/go.sum
… example/namedtracer/go.sum
… exporter/trace/jaeger/go.sum
… exporter/trace/stackdriver/go.sum
… go.sum
… sdk/trace/lrumap.go